requirements.txt
bambu-connect~=0.3.1
discord.py~=2.4.0
python-dotenv~=1.0.1